vst3: parse and normalise typed parameter text in double (#33)
getParameterValueForString parsed the text into a float and normalised it in
float arithmetic. Two float roundings (the parsed value, then the division by
the range) can land the normalised value one ulp away from where the plain
value sits, and when the host hands that normalised value back the plugin
receives a plain value one ulp off the text it was given.

Measured on DuskVerb 2 with the hosted renderer: "Decay Time=1.13125" typed
as text reached the DSP one ulp high and rendered -76 dB against the same
text typed into the JUCE build, while the same value set as a normalised
number (0.03125) was bit-identical. With the parse and the normalisation in
double the text route is bit-identical too, for 1.13125 and every other value
tried. CLAP's text_to_value already hands the host a plain double and was not
affected.

+++ daf/src/DafPluginVST3.cpp
@@ -1931,7 +1931,13 @@ public:
 
         const ScopedUTF8String input8(input);
 
-        float value;
+        // Parse and normalise in double. Doing it in float rounds twice (the
+        // parsed value, then the division by the range) and can land one ulp
+        // away from the plain value the text names once the host hands the
+        // normalised value back: "1.13125" s typed into DuskVerb 2 reached the
+        // DSP one ulp high, -76 dB against the same text in the JUCE build,
+        // while the same value set as a normalised number was bit-identical.
+        double value;
         if (fPlugin.getParameterHints(index) & kParameterIsInteger)
             value = std::atoi(input8);
         else
